Rename Projects to ProjectsAndFoldersLegacy
