+++ front/app/components/admin/ContentBuilder/Widgets/Container/index.tsx
@@ -4,7 +4,7 @@ import { Box } from '@citizenlab/cl2-component-library';
 import { UserComponent } from '@craftjs/core';
 
 import Events from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/Events';
-import Projects from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/Projects';
+import Projects from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/ProjectsAndFoldersLegacy';
 
 const Container: UserComponent = ({ children }) => {
   return (
